pyproject.toml
[project]
name = "bday-bot"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Casey Chance", email = "caseyjchance@gmail.com" }
]
dependencies = [
    "python-dotenv>=1.0.1",
    "discord.py>=2.3.2",
    "ipython>=8.12.3",
    "nltk>=3.8.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = [
    "black>=24.3.0",
]

[tool.hatch.metadata]
allow-direct-references = true
